Generate RIDMAE enumerated values
Thanks @AlyoshaVasilieva for the tip! svdtools can include enumerated
values when adding fields.
