+++ src/imxrt101/imxrt1011/mod.rs
@@ -42,10 +42,10 @@ pub mod usbnc;
 pub use super::instances::pgc;
 pub mod ccm;
 pub mod src;
+pub use super::instances::lpuart;
 pub use super::instances::romc;
 pub mod lpi2c;
 pub mod lpspi;
-pub mod lpuart;
 pub use super::instances::flexio1;
 pub mod gpio;
 pub mod pwm1;
