+++ src/imxrt101/imxrt1015/mod.rs
@@ -34,9 +34,9 @@ pub mod gpc;
 pub use super::instances::pgc;
 pub mod ccm;
 pub mod src;
-pub use super::instances::romc;
-pub mod lpuart;
 pub use super::instances::flexio1;
+pub use super::instances::lpuart;
+pub use super::instances::romc;
 pub mod gpio;
 pub mod tmr1;
 pub use super::instances::gpt;
