+++ src/imxrt101/peripherals/lpuart.rs
@@ -1,6 +1,8 @@
 #![allow(non_snake_case, non_upper_case_globals)]
 #![allow(non_camel_case_types)]
 //! LPUART
+//!
+//! Used by: imxrt1011, imxrt1015
 
 use crate::{RORegister, RWRegister};
 #[cfg(not(feature = "nosync"))]
@@ -296,6 +298,27 @@ pub mod BAUD {
         }
     }
 
+    /// Receiver Idle DMA Enable
+    pub mod RIDMAE {
+        /// Offset (20 bits)
+        pub const offset: u32 = 20;
+        /// Mask (1 bit: 1 << 20)
+        pub const mask: u32 = 1 << offset;
+        /// Read-only values (empty)
+        pub mod R {}
+        /// Write-only values (empty)
+        pub mod W {}
+        /// Read-write values
+        pub mod RW {
+
+            /// 0b0: DMA request disabled.
+            pub const RIDMAE_0: u32 = 0b0;
+
+            /// 0b1: DMA request enabled.
+            pub const RIDMAE_1: u32 = 0b1;
+        }
+    }
+
     /// Receiver Full DMA Enable
     pub mod RDMAE {
         /// Offset (21 bits)
@@ -503,27 +526,6 @@ pub mod BAUD {
 
             /// 0b1: Enables automatic address matching or data matching mode for MATCH\[MA1\].
             pub const MAEN1_1: u32 = 0b1;
-        }
-    }
-
-    /// Receiver Idle DMA Enable
-    pub mod RIDMAE {
-        /// Offset (20 bits)
-        pub const offset: u32 = 20;
-        /// Mask (1 bit: 1 << 20)
-        pub const mask: u32 = 1 << offset;
-        /// Read-only values (empty)
-        pub mod R {}
-        /// Write-only values (empty)
-        pub mod W {}
-        /// Read-write values
-        pub mod RW {
-
-            /// 0b0: DMA request disabled.
-            pub const RIDMAE_0: u32 = 0b0;
-
-            /// 0b1: DMA request enabled.
-            pub const RIDMAE_1: u32 = 0b1;
         }
     }
 }
@@ -2436,819 +2438,3 @@ impl ::core::ops::Deref for Instance {
 }
 
 unsafe impl Send for Instance {}
-
-/// Access functions for the LPUART1 peripheral instance
-pub mod LPUART1 {
-    use super::ResetValues;
-
-    #[cfg(not(feature = "nosync"))]
-    use super::Instance;
-
-    #[cfg(not(feature = "nosync"))]
-    const INSTANCE: Instance = Instance {
-        addr: 0x40184000,
-        _marker: ::core::marker::PhantomData,
-    };
-
-    /// Reset values for each field in LPUART1
-    pub const reset: ResetValues = ResetValues {
-        VERID: 0x04010003,
-        PARAM: 0x00000202,
-        GLOBAL: 0x00000000,
-        PINCFG: 0x00000000,
-        BAUD: 0x0F000004,
-        STAT: 0x00C00000,
-        CTRL: 0x00000000,
-        DATA: 0x00001000,
-        MATCH: 0x00000000,
-        MODIR: 0x00000000,
-        FIFO: 0x00C00011,
-        WATER: 0x00000000,
-    };
-
-    #[cfg(not(feature = "nosync"))]
-    #[allow(renamed_and_removed_lints)]
-    #[allow(private_no_mangle_statics)]
-    #[no_mangle]
-    static mut LPUART1_TAKEN: bool = false;
-
-    /// Safe access to LPUART1
-    ///
-    /// This function returns `Some(Instance)` if this instance is not
-    /// currently taken, and `None` if it is. This ensures that if you
-    /// do get `Some(Instance)`, you are ensured unique access to
-    /// the peripheral and there cannot be data races (unless other
-    /// code uses `unsafe`, of course). You can then pass the
-    /// `Instance` around to other functions as required. When you're
-    /// done with it, you can call `release(instance)` to return it.
-    ///
-    /// `Instance` itself dereferences to a `RegisterBlock`, which
-    /// provides access to the peripheral's registers.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn take() -> Option<Instance> {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART1_TAKEN {
-                None
-            } else {
-                LPUART1_TAKEN = true;
-                Some(INSTANCE)
-            }
-        })
-    }
-
-    /// Release exclusive access to LPUART1
-    ///
-    /// This function allows you to return an `Instance` so that it
-    /// is available to `take()` again. This function will panic if
-    /// you return a different `Instance` or if this instance is not
-    /// already taken.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn release(inst: Instance) {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART1_TAKEN && inst.addr == INSTANCE.addr {
-                LPUART1_TAKEN = false;
-            } else {
-                panic!("Released a peripheral which was not taken");
-            }
-        });
-    }
-
-    /// Unsafely steal LPUART1
-    ///
-    /// This function is similar to take() but forcibly takes the
-    /// Instance, marking it as taken irregardless of its previous
-    /// state.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub unsafe fn steal() -> Instance {
-        LPUART1_TAKEN = true;
-        INSTANCE
-    }
-}
-
-/// Raw pointer to LPUART1
-///
-/// Dereferencing this is unsafe because you are not ensured unique
-/// access to the peripheral, so you may encounter data races with
-/// other users of this peripheral. It is up to you to ensure you
-/// will not cause data races.
-///
-/// This constant is provided for ease of use in unsafe code: you can
-/// simply call for example `write_reg!(gpio, GPIOA, ODR, 1);`.
-pub const LPUART1: *const RegisterBlock = 0x40184000 as *const _;
-
-/// Access functions for the LPUART2 peripheral instance
-pub mod LPUART2 {
-    use super::ResetValues;
-
-    #[cfg(not(feature = "nosync"))]
-    use super::Instance;
-
-    #[cfg(not(feature = "nosync"))]
-    const INSTANCE: Instance = Instance {
-        addr: 0x40188000,
-        _marker: ::core::marker::PhantomData,
-    };
-
-    /// Reset values for each field in LPUART2
-    pub const reset: ResetValues = ResetValues {
-        VERID: 0x04010003,
-        PARAM: 0x00000202,
-        GLOBAL: 0x00000000,
-        PINCFG: 0x00000000,
-        BAUD: 0x0F000004,
-        STAT: 0x00C00000,
-        CTRL: 0x00000000,
-        DATA: 0x00001000,
-        MATCH: 0x00000000,
-        MODIR: 0x00000000,
-        FIFO: 0x00C00011,
-        WATER: 0x00000000,
-    };
-
-    #[cfg(not(feature = "nosync"))]
-    #[allow(renamed_and_removed_lints)]
-    #[allow(private_no_mangle_statics)]
-    #[no_mangle]
-    static mut LPUART2_TAKEN: bool = false;
-
-    /// Safe access to LPUART2
-    ///
-    /// This function returns `Some(Instance)` if this instance is not
-    /// currently taken, and `None` if it is. This ensures that if you
-    /// do get `Some(Instance)`, you are ensured unique access to
-    /// the peripheral and there cannot be data races (unless other
-    /// code uses `unsafe`, of course). You can then pass the
-    /// `Instance` around to other functions as required. When you're
-    /// done with it, you can call `release(instance)` to return it.
-    ///
-    /// `Instance` itself dereferences to a `RegisterBlock`, which
-    /// provides access to the peripheral's registers.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn take() -> Option<Instance> {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART2_TAKEN {
-                None
-            } else {
-                LPUART2_TAKEN = true;
-                Some(INSTANCE)
-            }
-        })
-    }
-
-    /// Release exclusive access to LPUART2
-    ///
-    /// This function allows you to return an `Instance` so that it
-    /// is available to `take()` again. This function will panic if
-    /// you return a different `Instance` or if this instance is not
-    /// already taken.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn release(inst: Instance) {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART2_TAKEN && inst.addr == INSTANCE.addr {
-                LPUART2_TAKEN = false;
-            } else {
-                panic!("Released a peripheral which was not taken");
-            }
-        });
-    }
-
-    /// Unsafely steal LPUART2
-    ///
-    /// This function is similar to take() but forcibly takes the
-    /// Instance, marking it as taken irregardless of its previous
-    /// state.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub unsafe fn steal() -> Instance {
-        LPUART2_TAKEN = true;
-        INSTANCE
-    }
-}
-
-/// Raw pointer to LPUART2
-///
-/// Dereferencing this is unsafe because you are not ensured unique
-/// access to the peripheral, so you may encounter data races with
-/// other users of this peripheral. It is up to you to ensure you
-/// will not cause data races.
-///
-/// This constant is provided for ease of use in unsafe code: you can
-/// simply call for example `write_reg!(gpio, GPIOA, ODR, 1);`.
-pub const LPUART2: *const RegisterBlock = 0x40188000 as *const _;
-
-/// Access functions for the LPUART3 peripheral instance
-pub mod LPUART3 {
-    use super::ResetValues;
-
-    #[cfg(not(feature = "nosync"))]
-    use super::Instance;
-
-    #[cfg(not(feature = "nosync"))]
-    const INSTANCE: Instance = Instance {
-        addr: 0x4018c000,
-        _marker: ::core::marker::PhantomData,
-    };
-
-    /// Reset values for each field in LPUART3
-    pub const reset: ResetValues = ResetValues {
-        VERID: 0x04010003,
-        PARAM: 0x00000202,
-        GLOBAL: 0x00000000,
-        PINCFG: 0x00000000,
-        BAUD: 0x0F000004,
-        STAT: 0x00C00000,
-        CTRL: 0x00000000,
-        DATA: 0x00001000,
-        MATCH: 0x00000000,
-        MODIR: 0x00000000,
-        FIFO: 0x00C00011,
-        WATER: 0x00000000,
-    };
-
-    #[cfg(not(feature = "nosync"))]
-    #[allow(renamed_and_removed_lints)]
-    #[allow(private_no_mangle_statics)]
-    #[no_mangle]
-    static mut LPUART3_TAKEN: bool = false;
-
-    /// Safe access to LPUART3
-    ///
-    /// This function returns `Some(Instance)` if this instance is not
-    /// currently taken, and `None` if it is. This ensures that if you
-    /// do get `Some(Instance)`, you are ensured unique access to
-    /// the peripheral and there cannot be data races (unless other
-    /// code uses `unsafe`, of course). You can then pass the
-    /// `Instance` around to other functions as required. When you're
-    /// done with it, you can call `release(instance)` to return it.
-    ///
-    /// `Instance` itself dereferences to a `RegisterBlock`, which
-    /// provides access to the peripheral's registers.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn take() -> Option<Instance> {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART3_TAKEN {
-                None
-            } else {
-                LPUART3_TAKEN = true;
-                Some(INSTANCE)
-            }
-        })
-    }
-
-    /// Release exclusive access to LPUART3
-    ///
-    /// This function allows you to return an `Instance` so that it
-    /// is available to `take()` again. This function will panic if
-    /// you return a different `Instance` or if this instance is not
-    /// already taken.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn release(inst: Instance) {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART3_TAKEN && inst.addr == INSTANCE.addr {
-                LPUART3_TAKEN = false;
-            } else {
-                panic!("Released a peripheral which was not taken");
-            }
-        });
-    }
-
-    /// Unsafely steal LPUART3
-    ///
-    /// This function is similar to take() but forcibly takes the
-    /// Instance, marking it as taken irregardless of its previous
-    /// state.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub unsafe fn steal() -> Instance {
-        LPUART3_TAKEN = true;
-        INSTANCE
-    }
-}
-
-/// Raw pointer to LPUART3
-///
-/// Dereferencing this is unsafe because you are not ensured unique
-/// access to the peripheral, so you may encounter data races with
-/// other users of this peripheral. It is up to you to ensure you
-/// will not cause data races.
-///
-/// This constant is provided for ease of use in unsafe code: you can
-/// simply call for example `write_reg!(gpio, GPIOA, ODR, 1);`.
-pub const LPUART3: *const RegisterBlock = 0x4018c000 as *const _;
-
-/// Access functions for the LPUART4 peripheral instance
-pub mod LPUART4 {
-    use super::ResetValues;
-
-    #[cfg(not(feature = "nosync"))]
-    use super::Instance;
-
-    #[cfg(not(feature = "nosync"))]
-    const INSTANCE: Instance = Instance {
-        addr: 0x40190000,
-        _marker: ::core::marker::PhantomData,
-    };
-
-    /// Reset values for each field in LPUART4
-    pub const reset: ResetValues = ResetValues {
-        VERID: 0x04010003,
-        PARAM: 0x00000202,
-        GLOBAL: 0x00000000,
-        PINCFG: 0x00000000,
-        BAUD: 0x0F000004,
-        STAT: 0x00C00000,
-        CTRL: 0x00000000,
-        DATA: 0x00001000,
-        MATCH: 0x00000000,
-        MODIR: 0x00000000,
-        FIFO: 0x00C00011,
-        WATER: 0x00000000,
-    };
-
-    #[cfg(not(feature = "nosync"))]
-    #[allow(renamed_and_removed_lints)]
-    #[allow(private_no_mangle_statics)]
-    #[no_mangle]
-    static mut LPUART4_TAKEN: bool = false;
-
-    /// Safe access to LPUART4
-    ///
-    /// This function returns `Some(Instance)` if this instance is not
-    /// currently taken, and `None` if it is. This ensures that if you
-    /// do get `Some(Instance)`, you are ensured unique access to
-    /// the peripheral and there cannot be data races (unless other
-    /// code uses `unsafe`, of course). You can then pass the
-    /// `Instance` around to other functions as required. When you're
-    /// done with it, you can call `release(instance)` to return it.
-    ///
-    /// `Instance` itself dereferences to a `RegisterBlock`, which
-    /// provides access to the peripheral's registers.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn take() -> Option<Instance> {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART4_TAKEN {
-                None
-            } else {
-                LPUART4_TAKEN = true;
-                Some(INSTANCE)
-            }
-        })
-    }
-
-    /// Release exclusive access to LPUART4
-    ///
-    /// This function allows you to return an `Instance` so that it
-    /// is available to `take()` again. This function will panic if
-    /// you return a different `Instance` or if this instance is not
-    /// already taken.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn release(inst: Instance) {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART4_TAKEN && inst.addr == INSTANCE.addr {
-                LPUART4_TAKEN = false;
-            } else {
-                panic!("Released a peripheral which was not taken");
-            }
-        });
-    }
-
-    /// Unsafely steal LPUART4
-    ///
-    /// This function is similar to take() but forcibly takes the
-    /// Instance, marking it as taken irregardless of its previous
-    /// state.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub unsafe fn steal() -> Instance {
-        LPUART4_TAKEN = true;
-        INSTANCE
-    }
-}
-
-/// Raw pointer to LPUART4
-///
-/// Dereferencing this is unsafe because you are not ensured unique
-/// access to the peripheral, so you may encounter data races with
-/// other users of this peripheral. It is up to you to ensure you
-/// will not cause data races.
-///
-/// This constant is provided for ease of use in unsafe code: you can
-/// simply call for example `write_reg!(gpio, GPIOA, ODR, 1);`.
-pub const LPUART4: *const RegisterBlock = 0x40190000 as *const _;
-
-/// Access functions for the LPUART5 peripheral instance
-pub mod LPUART5 {
-    use super::ResetValues;
-
-    #[cfg(not(feature = "nosync"))]
-    use super::Instance;
-
-    #[cfg(not(feature = "nosync"))]
-    const INSTANCE: Instance = Instance {
-        addr: 0x40194000,
-        _marker: ::core::marker::PhantomData,
-    };
-
-    /// Reset values for each field in LPUART5
-    pub const reset: ResetValues = ResetValues {
-        VERID: 0x04010003,
-        PARAM: 0x00000202,
-        GLOBAL: 0x00000000,
-        PINCFG: 0x00000000,
-        BAUD: 0x0F000004,
-        STAT: 0x00C00000,
-        CTRL: 0x00000000,
-        DATA: 0x00001000,
-        MATCH: 0x00000000,
-        MODIR: 0x00000000,
-        FIFO: 0x00C00011,
-        WATER: 0x00000000,
-    };
-
-    #[cfg(not(feature = "nosync"))]
-    #[allow(renamed_and_removed_lints)]
-    #[allow(private_no_mangle_statics)]
-    #[no_mangle]
-    static mut LPUART5_TAKEN: bool = false;
-
-    /// Safe access to LPUART5
-    ///
-    /// This function returns `Some(Instance)` if this instance is not
-    /// currently taken, and `None` if it is. This ensures that if you
-    /// do get `Some(Instance)`, you are ensured unique access to
-    /// the peripheral and there cannot be data races (unless other
-    /// code uses `unsafe`, of course). You can then pass the
-    /// `Instance` around to other functions as required. When you're
-    /// done with it, you can call `release(instance)` to return it.
-    ///
-    /// `Instance` itself dereferences to a `RegisterBlock`, which
-    /// provides access to the peripheral's registers.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn take() -> Option<Instance> {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART5_TAKEN {
-                None
-            } else {
-                LPUART5_TAKEN = true;
-                Some(INSTANCE)
-            }
-        })
-    }
-
-    /// Release exclusive access to LPUART5
-    ///
-    /// This function allows you to return an `Instance` so that it
-    /// is available to `take()` again. This function will panic if
-    /// you return a different `Instance` or if this instance is not
-    /// already taken.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn release(inst: Instance) {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART5_TAKEN && inst.addr == INSTANCE.addr {
-                LPUART5_TAKEN = false;
-            } else {
-                panic!("Released a peripheral which was not taken");
-            }
-        });
-    }
-
-    /// Unsafely steal LPUART5
-    ///
-    /// This function is similar to take() but forcibly takes the
-    /// Instance, marking it as taken irregardless of its previous
-    /// state.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub unsafe fn steal() -> Instance {
-        LPUART5_TAKEN = true;
-        INSTANCE
-    }
-}
-
-/// Raw pointer to LPUART5
-///
-/// Dereferencing this is unsafe because you are not ensured unique
-/// access to the peripheral, so you may encounter data races with
-/// other users of this peripheral. It is up to you to ensure you
-/// will not cause data races.
-///
-/// This constant is provided for ease of use in unsafe code: you can
-/// simply call for example `write_reg!(gpio, GPIOA, ODR, 1);`.
-pub const LPUART5: *const RegisterBlock = 0x40194000 as *const _;
-
-/// Access functions for the LPUART6 peripheral instance
-pub mod LPUART6 {
-    use super::ResetValues;
-
-    #[cfg(not(feature = "nosync"))]
-    use super::Instance;
-
-    #[cfg(not(feature = "nosync"))]
-    const INSTANCE: Instance = Instance {
-        addr: 0x40198000,
-        _marker: ::core::marker::PhantomData,
-    };
-
-    /// Reset values for each field in LPUART6
-    pub const reset: ResetValues = ResetValues {
-        VERID: 0x04010003,
-        PARAM: 0x00000202,
-        GLOBAL: 0x00000000,
-        PINCFG: 0x00000000,
-        BAUD: 0x0F000004,
-        STAT: 0x00C00000,
-        CTRL: 0x00000000,
-        DATA: 0x00001000,
-        MATCH: 0x00000000,
-        MODIR: 0x00000000,
-        FIFO: 0x00C00011,
-        WATER: 0x00000000,
-    };
-
-    #[cfg(not(feature = "nosync"))]
-    #[allow(renamed_and_removed_lints)]
-    #[allow(private_no_mangle_statics)]
-    #[no_mangle]
-    static mut LPUART6_TAKEN: bool = false;
-
-    /// Safe access to LPUART6
-    ///
-    /// This function returns `Some(Instance)` if this instance is not
-    /// currently taken, and `None` if it is. This ensures that if you
-    /// do get `Some(Instance)`, you are ensured unique access to
-    /// the peripheral and there cannot be data races (unless other
-    /// code uses `unsafe`, of course). You can then pass the
-    /// `Instance` around to other functions as required. When you're
-    /// done with it, you can call `release(instance)` to return it.
-    ///
-    /// `Instance` itself dereferences to a `RegisterBlock`, which
-    /// provides access to the peripheral's registers.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn take() -> Option<Instance> {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART6_TAKEN {
-                None
-            } else {
-                LPUART6_TAKEN = true;
-                Some(INSTANCE)
-            }
-        })
-    }
-
-    /// Release exclusive access to LPUART6
-    ///
-    /// This function allows you to return an `Instance` so that it
-    /// is available to `take()` again. This function will panic if
-    /// you return a different `Instance` or if this instance is not
-    /// already taken.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn release(inst: Instance) {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART6_TAKEN && inst.addr == INSTANCE.addr {
-                LPUART6_TAKEN = false;
-            } else {
-                panic!("Released a peripheral which was not taken");
-            }
-        });
-    }
-
-    /// Unsafely steal LPUART6
-    ///
-    /// This function is similar to take() but forcibly takes the
-    /// Instance, marking it as taken irregardless of its previous
-    /// state.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub unsafe fn steal() -> Instance {
-        LPUART6_TAKEN = true;
-        INSTANCE
-    }
-}
-
-/// Raw pointer to LPUART6
-///
-/// Dereferencing this is unsafe because you are not ensured unique
-/// access to the peripheral, so you may encounter data races with
-/// other users of this peripheral. It is up to you to ensure you
-/// will not cause data races.
-///
-/// This constant is provided for ease of use in unsafe code: you can
-/// simply call for example `write_reg!(gpio, GPIOA, ODR, 1);`.
-pub const LPUART6: *const RegisterBlock = 0x40198000 as *const _;
-
-/// Access functions for the LPUART7 peripheral instance
-pub mod LPUART7 {
-    use super::ResetValues;
-
-    #[cfg(not(feature = "nosync"))]
-    use super::Instance;
-
-    #[cfg(not(feature = "nosync"))]
-    const INSTANCE: Instance = Instance {
-        addr: 0x4019c000,
-        _marker: ::core::marker::PhantomData,
-    };
-
-    /// Reset values for each field in LPUART7
-    pub const reset: ResetValues = ResetValues {
-        VERID: 0x04010003,
-        PARAM: 0x00000202,
-        GLOBAL: 0x00000000,
-        PINCFG: 0x00000000,
-        BAUD: 0x0F000004,
-        STAT: 0x00C00000,
-        CTRL: 0x00000000,
-        DATA: 0x00001000,
-        MATCH: 0x00000000,
-        MODIR: 0x00000000,
-        FIFO: 0x00C00011,
-        WATER: 0x00000000,
-    };
-
-    #[cfg(not(feature = "nosync"))]
-    #[allow(renamed_and_removed_lints)]
-    #[allow(private_no_mangle_statics)]
-    #[no_mangle]
-    static mut LPUART7_TAKEN: bool = false;
-
-    /// Safe access to LPUART7
-    ///
-    /// This function returns `Some(Instance)` if this instance is not
-    /// currently taken, and `None` if it is. This ensures that if you
-    /// do get `Some(Instance)`, you are ensured unique access to
-    /// the peripheral and there cannot be data races (unless other
-    /// code uses `unsafe`, of course). You can then pass the
-    /// `Instance` around to other functions as required. When you're
-    /// done with it, you can call `release(instance)` to return it.
-    ///
-    /// `Instance` itself dereferences to a `RegisterBlock`, which
-    /// provides access to the peripheral's registers.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn take() -> Option<Instance> {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART7_TAKEN {
-                None
-            } else {
-                LPUART7_TAKEN = true;
-                Some(INSTANCE)
-            }
-        })
-    }
-
-    /// Release exclusive access to LPUART7
-    ///
-    /// This function allows you to return an `Instance` so that it
-    /// is available to `take()` again. This function will panic if
-    /// you return a different `Instance` or if this instance is not
-    /// already taken.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn release(inst: Instance) {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART7_TAKEN && inst.addr == INSTANCE.addr {
-                LPUART7_TAKEN = false;
-            } else {
-                panic!("Released a peripheral which was not taken");
-            }
-        });
-    }
-
-    /// Unsafely steal LPUART7
-    ///
-    /// This function is similar to take() but forcibly takes the
-    /// Instance, marking it as taken irregardless of its previous
-    /// state.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub unsafe fn steal() -> Instance {
-        LPUART7_TAKEN = true;
-        INSTANCE
-    }
-}
-
-/// Raw pointer to LPUART7
-///
-/// Dereferencing this is unsafe because you are not ensured unique
-/// access to the peripheral, so you may encounter data races with
-/// other users of this peripheral. It is up to you to ensure you
-/// will not cause data races.
-///
-/// This constant is provided for ease of use in unsafe code: you can
-/// simply call for example `write_reg!(gpio, GPIOA, ODR, 1);`.
-pub const LPUART7: *const RegisterBlock = 0x4019c000 as *const _;
-
-/// Access functions for the LPUART8 peripheral instance
-pub mod LPUART8 {
-    use super::ResetValues;
-
-    #[cfg(not(feature = "nosync"))]
-    use super::Instance;
-
-    #[cfg(not(feature = "nosync"))]
-    const INSTANCE: Instance = Instance {
-        addr: 0x401a0000,
-        _marker: ::core::marker::PhantomData,
-    };
-
-    /// Reset values for each field in LPUART8
-    pub const reset: ResetValues = ResetValues {
-        VERID: 0x04010003,
-        PARAM: 0x00000202,
-        GLOBAL: 0x00000000,
-        PINCFG: 0x00000000,
-        BAUD: 0x0F000004,
-        STAT: 0x00C00000,
-        CTRL: 0x00000000,
-        DATA: 0x00001000,
-        MATCH: 0x00000000,
-        MODIR: 0x00000000,
-        FIFO: 0x00C00011,
-        WATER: 0x00000000,
-    };
-
-    #[cfg(not(feature = "nosync"))]
-    #[allow(renamed_and_removed_lints)]
-    #[allow(private_no_mangle_statics)]
-    #[no_mangle]
-    static mut LPUART8_TAKEN: bool = false;
-
-    /// Safe access to LPUART8
-    ///
-    /// This function returns `Some(Instance)` if this instance is not
-    /// currently taken, and `None` if it is. This ensures that if you
-    /// do get `Some(Instance)`, you are ensured unique access to
-    /// the peripheral and there cannot be data races (unless other
-    /// code uses `unsafe`, of course). You can then pass the
-    /// `Instance` around to other functions as required. When you're
-    /// done with it, you can call `release(instance)` to return it.
-    ///
-    /// `Instance` itself dereferences to a `RegisterBlock`, which
-    /// provides access to the peripheral's registers.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn take() -> Option<Instance> {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART8_TAKEN {
-                None
-            } else {
-                LPUART8_TAKEN = true;
-                Some(INSTANCE)
-            }
-        })
-    }
-
-    /// Release exclusive access to LPUART8
-    ///
-    /// This function allows you to return an `Instance` so that it
-    /// is available to `take()` again. This function will panic if
-    /// you return a different `Instance` or if this instance is not
-    /// already taken.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub fn release(inst: Instance) {
-        external_cortex_m::interrupt::free(|_| unsafe {
-            if LPUART8_TAKEN && inst.addr == INSTANCE.addr {
-                LPUART8_TAKEN = false;
-            } else {
-                panic!("Released a peripheral which was not taken");
-            }
-        });
-    }
-
-    /// Unsafely steal LPUART8
-    ///
-    /// This function is similar to take() but forcibly takes the
-    /// Instance, marking it as taken irregardless of its previous
-    /// state.
-    #[cfg(not(feature = "nosync"))]
-    #[inline]
-    pub unsafe fn steal() -> Instance {
-        LPUART8_TAKEN = true;
-        INSTANCE
-    }
-}
-
-/// Raw pointer to LPUART8
-///
-/// Dereferencing this is unsafe because you are not ensured unique
-/// access to the peripheral, so you may encounter data races with
-/// other users of this peripheral. It is up to you to ensure you
-/// will not cause data races.
-///
-/// This constant is provided for ease of use in unsafe code: you can
-/// simply call for example `write_reg!(gpio, GPIOA, ODR, 1);`.
-pub const LPUART8: *const RegisterBlock = 0x401a0000 as *const _;
