+++ src/imxrt101/peripherals/mod.rs
@@ -56,6 +56,9 @@ pub mod pgc;
 pub mod romc;
 
 #[cfg(any(feature = "doc", feature = "imxrt1011", feature = "imxrt1015"))]
+pub mod lpuart;
+
+#[cfg(any(feature = "doc", feature = "imxrt1011", feature = "imxrt1015"))]
 pub mod lpi2c;
 
 #[cfg(any(feature = "doc", feature = "imxrt1011", feature = "imxrt1015"))]
